Restricted conditions (= relaxed disambiguation)
Made a set LIST farlige-Pl-A = ("gammel" A Pl Def) ... for adjectives like de gamle, unge, gjerrige, ... (A-as-N, even when disambiguated as A)

Restricted the SELECT A Pl for "bare" to *1C N Pl (earlier only *1), N Pl could really be anything, at least a verb.

Added farlige-Pl-A set to rule #2414 (selecting N Pl if A Pl to the left, but not for thesee dangerous A).