Rule revisions
Define Clt as tag. The tag is relevant to the restriction msyn-np-demindef-demdef , since the indef pl nouns shouild not be Def when the A Def to the left as a cliti994117b2

Ad NOT -1 Superl for msyn-agr-adjsg-adjpl since *mange organisert aktiviteter but flest mulig samer.

Added C to the demand on 1 N Neu Sg Indef for msyn-agr-detmsc-detneu.

Added the start to a rule &real-cp-null-som, but am not sure how to do it, since the goal is to insert a new word.